Update build.gradle
with the added code it also generates *-sources.jar and *-docs.jar . I am new with Bintray but I had the surprise of package rejection because of missing sources & docs. That code is from http://stackoverflow.com/questions/11474729/how-to-build-sources-jar-with-gradle and it was given to be by someone from support, telling me to add the missing jars.